get_language_model tests and adaptive model tests
